# apps/api/src/swagger_output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Medialit API",
    "description": "Easy file uploads for serverless apps",
    "version": "0.3.0",
    "termsOfService": "https://medialit.cloud/p/terms",
    "contact": {
      "name": "Medialit Support",
      "url": "https://medialit.cloud"
    },
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    }
  },
  "servers": [
    {
      "url": "{protocol}://{host}",
      "description": "API Server",
      "variables": {
        "protocol": {
          "default": "https",
          "enum": [
            "https",
            "http"
          ]
        },
        "host": {
          "default": "api.medialit.cloud"
        }
      }
    }
  ],
...
  "paths": {
...
    "/oauth/token": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Bad Request"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/oauth/register": {
      "post": {
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Bad Request"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-medialit-apikey"
      },
      "signatureAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-medialit-signature"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Unauthorized"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      }
    }
  }
}